+++ Week 6/WeatherMonitoring/WeatherMonitoring/currentValuesObserver.cs
@@ -12,7 +12,7 @@ namespace WeatherMonitoring
         public currentValuesObserver(ListBox displayBox, WeatherMonitoringSubject subject) : base(displayBox,subject)
         {}
 
-        public override void update(int currTemp, int currHumidity, int currBaroPressure)
+        public override void update(double currTemp, double currHumidity, double currBaroPressure)
         {
             currtemp = currTemp;
             currhumidity = currHumidity;
